Adding DPI parameter to PDF tests and print test result as a method
Adding the option for specifying the DPI as parameter for PDF tests.
This parameter fix some bugs for example:
1) Stream truncated! (Block truncated)]
2) Java heap space - it simply allows to reduce the converted image
size.

In addition this commit includes a standard function for printing the
test result to console.

Last but not least in PDF test, the conversion process of the pdf into
image got shorter